+++ client/src/pages/CreateSurvey/Questions/Question.jsx
@@ -126,7 +126,6 @@ const Question = ({ questionIndex }) => {
             />
           </Box>
           {/*  COMMENTS ENABLED CHECKBOX */}
-          {/* TODO make comments enabled work */}
           <Box>
             <FormControlLabel
               control={
@@ -157,41 +156,4 @@ Question.propTypes = {
   questionIndex: PropTypes.number.isRequired,
 };
 
-const QuestionsList = () => {
-  const dispatch = useDispatch();
-  const handleNewQuestionClick = () => {
-    dispatch({ type: 'NEW_QUESTION' });
-  };
-  const { questions } = useSelector((state) => state.createSurveyReducer);
-
-  if (questions.length < 1) {
-    handleNewQuestionClick();
-  }
-
-  return (
-    <Box mx={4} my={2} display='flex' flexDirection='column'>
-      {questions.map((question, questionIndex) => {
-        return (
-          <Question
-            // eslint-disable-next-line react/no-array-index-key
-            key={questionIndex}
-            question={question}
-            questionIndex={questionIndex}
-          />
-        );
-      })}
-      <Box alignSelf='flex-start' my={2}>
-        <Button
-          type='button'
-          variant='outlined'
-          color='secondary'
-          onClick={handleNewQuestionClick}
-        >
-          Add question
-        </Button>
-      </Box>
-    </Box>
-  );
-};
-
-export default QuestionsList;
+export default Question;
