+++ client/src/pages/CreateSurvey/Questions/Questions.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Provider } from 'react-redux';
 import { render, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
-import Questions from './Questions';
+import Questions from './QuestionsList';
 import store from '../../../store/index';
 
 describe('Questions component works as expected', () => {
